fix/Modify post API due to API changes

+++ controllers/posts.controller.js
@@ -7,42 +7,43 @@ class PostController {
 
   getLocationPosts = async (req, res, next) => {
     try {
-      const { postLocation } = req.query;
-      const posts = await this.postService.getLocationPosts(postLocation);
+      const { postLocation1, postLocation2 } = req.query;
+      const posts = await this.postService.getLocationPosts(
+        postLocation1,
+        postLocation2,
+      );
       res.status(200).json({ posts });
     } catch (err) {
-      if (err.isJoi === true)
-        return res.status(400).json({ errorMsg: '유효성 검사 에러' });
-      res
-        .status(err.code || 500)
-        .json({ errorMsg: err.msg || '알수없는 에러' });
+      next(err);
     }
   };
 
   createPost = async (req, res, next) => {
     try {
       const userId = res.locals;
-      const { postLocation } = req.query;
-      const { title, content } = req.body;
+      const { title, content, postLocation1, postLocation2 } = req.body;
       if (req.file) {
         const postImage = req.file.location;
         await this.postService.createPost(
           title,
           content,
-          postLocation,
+          postLocation1,
+          postLocation2,
           userId,
           postImage,
         );
       } else
-        await this.postService.createPost(title, content, postLocation, userId);
+        await this.postService.createPost(
+          title,
+          content,
+          postLocation1,
+          postLocation2,
+          userId,
+        );
       res.status(200).json({ msg: '작성 성공' });
     } catch (err) {
-      await this.postService.deleteS3Image(req.file.key);
-      if (err.isJoi === true)
-        return res.status(400).json({ errorMsg: '유효성 검사 에러' });
-      res
-        .status(err.code || 500)
-        .json({ errorMsg: err.msg || '알수없는 에러' });
+      if (req.file) await this.postService.deleteS3Image(req.file.key);
+      next(err);
     }
   };
 
@@ -52,11 +53,7 @@ class PostController {
       const post = await this.postService.getDetailPost(postId);
       res.status(200).json({ post });
     } catch (err) {
-      if (err.isJoi === true)
-        return res.status(400).json({ errorMsg: '유효성 검사 에러' });
-      res
-        .status(err.code || 500)
-        .json({ errorMsg: err.msg || '알수없는 에러' });
+      next(err);
     }
   };
 
@@ -67,11 +64,7 @@ class PostController {
       const post = await this.postService.getPreviousPost(postId, userId);
       res.status(200).json({ post });
     } catch (err) {
-      if (err.isJoi === true)
-        return res.status(400).json({ errorMsg: '유효성 검사 에러' });
-      res
-        .status(err.code || 500)
-        .json({ errorMsg: err.msg || '알수없는 에러' });
+      next(err);
     }
   };
 
@@ -92,12 +85,8 @@ class PostController {
       } else await this.postService.updatePost(postId, title, content, userId);
       res.status(200).json({ msg: '수정 완료' });
     } catch (err) {
-      await this.postService.deleteS3Image(req.file.key);
-      if (err.isJoi === true)
-        return res.status(400).json({ errorMsg: '유효성 검사 에러' });
-      res
-        .status(err.code || 500)
-        .json({ errorMsg: err.msg || '알수없는 에러' });
+      if (req.file) await this.postService.deleteS3Image(req.file.key);
+      next(err);
     }
   };
 
@@ -108,11 +97,17 @@ class PostController {
       await this.postService.deletePost(postId, userId);
       res.status(200).json({ msg: '삭제 완료' });
     } catch (err) {
-      if (err.isJoi === true)
-        return res.status(400).json({ errorMsg: '유효성 검사 에러' });
-      res
-        .status(err.code || 500)
-        .json({ errorMsg: err.msg || '알수없는 에러' });
+      next(err);
+    }
+  };
+
+  getMyPost = async (req, res, next) => {
+    try {
+      const userId = res.locals;
+      const myposts = await this.postService.getMyPost(userId);
+      res.status(200).json({ myposts });
+    } catch (err) {
+      next(err);
     }
   };
 }
